Revert "Block Bindings: Refactor e2e tests (#65526)"
This reverts commit b3413730c71fc058118f02c94143ab5a087695ce.
