--- test/gutenberg-test-themes/block-bindings/index.php
@@ -1,9 +1,0 @@
-<?php
-/**
- * Theme index.php file.
- *
- * @package Gutenberg
- */
-
-// Silence is golden.
-return;
