--- test/gutenberg-test-themes/block-bindings/style.css
@@ -1,15 +1,0 @@
-/*
-Theme Name: Block Bindings
-Theme URI: https://github.com/wordpress/gutenberg/
-Author: the WordPress team
-Description: Block bindings test theme.
-Requires at least: 6.7
-Tested up to: 6.7
-Requires PHP: 7.2
-Version: 1.0
-License: GNU General Public License v2 or later
-License URI: http://www.gnu.org/licenses/gpl-2.0.html
-Text Domain: block-bindings
-Block bindings WordPress Theme, (C) 2024 WordPress.org
-Block bindings is distributed under the terms of the GNU GPL.
-*/
