--- test/gutenberg-test-themes/block-bindings/templates/custom-template.html
@@ -1,3 +1,0 @@
-<!-- wp:heading -->
-<h2 class="wp-block-heading">Custom template</h2>
-<!-- /wp:heading -->
