--- test/gutenberg-test-themes/block-bindings/templates/index.html
@@ -1,8 +1,0 @@
-<!-- wp:query {"queryId":1,"query":{"offset":0,"postType":"post","order":"desc","orderBy":"date","author":"","search":"","sticky":""}} -->
-<div class="wp-block-query">
-	<!-- wp:post-template -->
-	<!-- wp:post-title {"isLink":true} /-->
-	<!-- wp:post-excerpt /-->
-	<!-- /wp:post-template -->
-</div>
-<!-- /wp:query -->
