--- test/gutenberg-test-themes/block-bindings/templates/single-movie.html
@@ -1,2 +1,0 @@
-<!-- wp:post-title /-->
-<!-- wp:post-content {"layout":{"inherit":true}} /-->
